align: add alloc/free implementation based on operator new
default and fallback implementation map `aligned_alloc` and
`aligned_free` to malloc/free and the platform equivalents.

however there are certain advantages when using the global
operator `new`/`delete`, e.g. for applications that provide
custom implementations of these operators. we therefore add a new
implementation based on the c++ operators. when c++17 is available
we directly map to alignment-aware operators, otherwise we use the
trick of re-aligning un-aligned memory